Apps: Extract BinaryXmlFormatting from ManifestTextRenderer
Pulls the pure formatting helpers (qualifiedName, attributeName, formatAttributeValue, formatReference, formatColor, formatComplex, escapeXml) plus the ResourceRefResolver fun interface into a shared internal object. Renderer now consumes them; all behaviour preserved.

Sets the stage for ManifestSectionVisitor to share the same formatting machinery without duplicating renderer state.
